// packages/grafana-data/src/types/flot.ts
export interface FlotGradient {
  colors: Array<
    | {
        opacity?: number;
        brightness?: number;
      }
    | string
  >;
}
// ...
export interface DisplayOptions {
  show?: boolean;
  lineWidth?: number;
  fill?: boolean | number;
}

export interface FlotLineOptions extends DisplayOptions {
  zero?: boolean;
  steps?: boolean;
  fillColor?: null | false | string;
}

export interface FlotBarOptions extends DisplayOptions {
  zero?: boolean;
  barWidth?: number; // in units of x-axis
  align?: 'left' | 'right' | 'center';
  horizontal?: boolean;
  fillColor?: null | false | string | FlotGradient;
}

export interface FlotPointOptions extends DisplayOptions {
  radius?: number;
  symbol?: 'circle' | ((ctx: any, x: number, y: number, radius: number, shadow: boolean) => void);
  fillColor?: null | false | string;
}

export interface FlotSeries {
  data: Array<[number, number]>;

  color?: string | number;
  label?: string;
  lines?: FlotLineOptions;
  bars?: FlotBarOptions;
  points?: FlotPointOptions;
  xaxis?: number; // ID of x-axis to plot series against
  yaxis?: number; // ID of y-axis to plot series against
  clickable?: boolean;
  hoverable?: boolean;
  shadowSize?: number;
  highlightColor?: string | number;

  // Stack options
  stack?: null | boolean | number | string;
}